рефакторинг интерфейсов

+++ Common/StmDogovorPro.Domain/ExcelBase/AnnualVolume.cs
@@ -4,12 +4,12 @@ namespace StmDogovorPro.Domain.ExcelBase
 {
     public class AnnualVolume : IAnnualVolume
     {
-        public int RRC { get; set; }
-        public int TD { get; set; }
-        public decimal FK { get; set; }
-        public decimal NAS { get; set; }
-        public decimal SGA { get; set; }
-        public decimal HF { get; set; }
-        public decimal Total { get; set; }
+        public double RRC { get; set; }
+        public double TD { get; set; }
+        public double FK { get; set; }
+        public double NAS { get; set; }
+        public double SGA { get; set; }
+        public double HF { get; set; }
+        public double Total { get; set; }
     }
 }
