+++ Common/StmDogovorPro.Domain/ExcelBase/RCElement.cs
@@ -4,10 +4,10 @@ namespace StmDogovorPro.Domain.ExcelBase
 {
     public class RCElement : IRCElement
     {
-        public int TD { get; set; }
-        public decimal FK { get; set; }
-        public decimal NAS { get; set; }
-        public decimal SGA { get; set; }
-        public decimal HF { get; set; }
+        public double TD { get; set; }
+        public double FK { get; set; }
+        public double NAS { get; set; }
+        public double SGA { get; set; }
+        public double HF { get; set; }
     }
 }
